change default class name from Foo to MyClass in tests
