+++ tests/Generator/Fixtures/Basic/Input/MyClass.json
