+++ tests/Generator/Fixtures/DefaultValueAsOptional/Output/MyClass.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Ns\NoGetters;
+namespace Ns\DefaultValueAsOptional;
 
-class Foo
+class MyClass
 {
     /**
      * Schema used to validate input for creating instances of this class
@@ -13,45 +13,85 @@ class Foo
      */
     private static array $schema = [
         'required' => [
-            'foo',
+            
         ],
         'properties' => [
-            'foo' => [
-                'type' => 'string',
+            'limit' => [
+                'type' => 'integer',
+                'default' => 10000,
+                'minimum' => 1,
+            ],
+            'skip' => [
+                'type' => 'integer',
+                'default' => 0,
             ],
         ],
     ];
 
     /**
-     * @var string
+     * @var int
      */
-    public string $foo;
+    private int $limit = 10000;
 
     /**
-     * @param string $foo
+     * @var int
      */
-    public function __construct(string $foo)
+    private int $skip = 0;
+
+    /**
+     * @return int
+     */
+    public function getLimit() : int
     {
-        $this->foo = $foo;
+        return $this->limit;
     }
 
     /**
-     * @param string $foo
+     * @return int
+     */
+    public function getSkip() : int
+    {
+        return $this->skip;
+    }
+
+    /**
+     * @param int $limit
      * @return self
      * @param bool $validate
      */
-    public function withFoo(string $foo, bool $validate = true) : self
+    public function withLimit(int $limit, bool $validate = true) : self
     {
         if ($validate) {
             $validator = new \JsonSchema\Validator();
-            $validator->validate($foo, self::$schema['properties']['foo']);
+            $validator->validate($limit, self::$schema['properties']['limit']);
             if (!$validator->isValid()) {
                 throw new \InvalidArgumentException($validator->getErrors()[0]['message']);
             }
         }
 
         $clone = clone $this;
-        $clone->foo = $foo;
+        $clone->limit = $limit;
+
+        return $clone;
+    }
+
+    /**
+     * @param int $skip
+     * @return self
+     * @param bool $validate
+     */
+    public function withSkip(int $skip, bool $validate = true) : self
+    {
+        if ($validate) {
+            $validator = new \JsonSchema\Validator();
+            $validator->validate($skip, self::$schema['properties']['skip']);
+            if (!$validator->isValid()) {
+                throw new \InvalidArgumentException($validator->getErrors()[0]['message']);
+            }
+        }
+
+        $clone = clone $this;
+        $clone->skip = $skip;
 
         return $clone;
     }
@@ -61,20 +101,28 @@ class Foo
      *
      * @param array|object $input Input data
      * @param bool $validate Set this to false to skip validation; use at own risk
-     * @return Foo Created instance
+     * @return MyClass Created instance
      * @throws \InvalidArgumentException
      */
-    public static function buildFromInput(array|object $input, bool $validate = true) : Foo
+    public static function buildFromInput(array|object $input, bool $validate = true) : MyClass
     {
         $input = is_array($input) ? \JsonSchema\Validator::arrayToObjectRecursive($input) : $input;
         if ($validate) {
             static::validateInput($input);
         }
 
-        $foo = $input->{'foo'};
+        $limit = 10000;
+        if (isset($input->{'limit'})) {
+            $limit = (int)($input->{'limit'});
+        }
+        $skip = 0;
+        if (isset($input->{'skip'})) {
+            $skip = (int)($input->{'skip'});
+        }
 
-        $obj = new self($foo);
-
+        $obj = new self();
+        $obj->limit = $limit;
+        $obj->skip = $skip;
         return $obj;
     }
 
@@ -86,7 +134,8 @@ class Foo
     public function toArray() : array
     {
         $output = [];
-        $output['foo'] = $this->foo;
+        $output['limit'] = $this->limit;
+        $output['skip'] = $this->skip;
 
         return $output;
     }
