+++ tests/Generator/Fixtures/EnumConsistent/Output/MyClass.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Ns\EnumConsistent;
 
-enum Foo: string {
+enum MyClass: string {
     case VALUE_FOO = 'Foo';
     case VALUE_1FOO = '1Foo';
 }
