+++ tests/Generator/Fixtures/SuperglobalCollision/Output/MyClass.php
@@ -2,55 +2,64 @@
 
 declare(strict_types=1);
 
-namespace Ns\SingleLineSchema;
+namespace Ns\SuperglobalCollision;
 
-class Foo
+class MyClass
 {
     /**
      * Schema used to validate input for creating instances of this class
      *
      * @var array
      */
-    private static array $schema = ['required' => ['foo'], 'properties' => ['foo' => ['type' => 'string']]];
+    private static array $schema = [
+        'required' => [
+            'files',
+        ],
+        'properties' => [
+            'files' => [
+                'type' => 'string',
+            ],
+        ],
+    ];
 
     /**
      * @var string
      */
-    private string $foo;
+    private string $files;
 
     /**
-     * @param string $foo
+     * @param string $files
      */
-    public function __construct(string $foo)
+    public function __construct(string $files)
     {
-        $this->foo = $foo;
+        $this->files = $files;
     }
 
     /**
      * @return string
      */
-    public function getFoo() : string
+    public function getFiles() : string
     {
-        return $this->foo;
+        return $this->files;
     }
 
     /**
-     * @param string $foo
+     * @param string $files
      * @return self
      * @param bool $validate
      */
-    public function withFoo(string $foo, bool $validate = true) : self
+    public function withFiles(string $files, bool $validate = true) : self
     {
         if ($validate) {
             $validator = new \JsonSchema\Validator();
-            $validator->validate($foo, self::$schema['properties']['foo']);
+            $validator->validate($files, self::$schema['properties']['files']);
             if (!$validator->isValid()) {
                 throw new \InvalidArgumentException($validator->getErrors()[0]['message']);
             }
         }
 
         $clone = clone $this;
-        $clone->foo = $foo;
+        $clone->files = $files;
 
         return $clone;
     }
@@ -60,19 +69,19 @@ class Foo
      *
      * @param array|object $input Input data
      * @param bool $validate Set this to false to skip validation; use at own risk
-     * @return Foo Created instance
+     * @return MyClass Created instance
      * @throws \InvalidArgumentException
      */
-    public static function buildFromInput(array|object $input, bool $validate = true) : Foo
+    public static function buildFromInput(array|object $input, bool $validate = true) : MyClass
     {
         $input = is_array($input) ? \JsonSchema\Validator::arrayToObjectRecursive($input) : $input;
         if ($validate) {
             static::validateInput($input);
         }
 
-        $foo = $input->{'foo'};
+        $files = $input->{'files'};
 
-        $obj = new self($foo);
+        $obj = new self($files);
 
         return $obj;
     }
@@ -85,7 +94,7 @@ class Foo
     public function toArray() : array
     {
         $output = [];
-        $output['foo'] = $this->foo;
+        $output['files'] = $this->files;
 
         return $output;
     }
